+++ tests/kafkatest/tests/streams/utils/util.py
@@ -1,0 +1,29 @@
+# Copyright 2015 Confluent Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+def verify_running(processor, message):
+    node = processor.node
+    with node.account.monitor_log(processor.STDOUT_FILE) as monitor:
+        processor.start()
+        monitor.wait_until(message,
+                           timeout_sec=60,
+                           err_msg="Never saw '%s' message " % message + str(processor.node.account))
+
+def verify_stopped(processor, message):
+    node = processor.node
+    with node.account.monitor_log(processor.STDOUT_FILE) as monitor:
+        processor.stop()
+        monitor.wait_until('Static membership test closed',
+                           timeout_sec=60,
+                           err_msg="'%s' message " % message + str(processor.node.account))
